# feat(indexes): index (->>) and lower(->>) on secondary keys