--- a/.travis.yml
+++ b/.travis.yml
@@ -7,4 +7,4 @@ before_script:
     - yes | python setup.py install
     - python bin/destroy_and_setup_psqlgraph.py
     - pip freeze
-script: "nosetests -v"
+script: "py.test -v"