We can get approximately two orders of magnitude speedup on certain frequently used classes of queries by adding indexes on the text representation of the json fields defined as secondary keys in the dictionary.

### Before
```
SELECT count(*) FROM node_case WHERE (lower(node_case._props ->> 'submitter_id') LIKE 'tcga-c9' || '%%');
Time: 12.159 ms

%time g.nodes().filter(Node._props['submitter_id'].astext.startswith('TCGA-C9-A4')).count()
Wall time: 3.94 s
```

### After
```
SELECT count(*) FROM node_case WHERE (lower(node_case._props ->> 'submitter_id') LIKE 'tcga-c9' || '%%');
Time: 0.604 ms

 %time g.nodes().filter(Node._props['submitter_id'].astext.startswith('TCGA-C9-A4')).count()
CPU times: user 17 ms, sys: 3.78 ms, total: 20.8 ms
Wall time: 39.2 ms
```